+++ resources/lib/objects/__init__.py
@@ -1,1 +1,2 @@
 # Dummy file to make this directory a package.
+from movies import Movies
